+++ src/Exceptions/Exception.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Php\Support\Exceptions;
 
+use Php\Support\Traits\Thrower;
 use Throwable;
 
 /**
@@ -13,6 +14,8 @@ use Throwable;
  */
 class Exception extends \Exception
 {
+    use Thrower;
+
     /**
      * Exception constructor.
      *